Added scripts folder and placeholder file
The scripts folder is a place for us to store any scripts that can be run independently from the rest of the system using npm run 'scriptname' aliases (when setup in package.json).

For example, to pull in the latest data from the student data system for eleventy, we run the command "npm run update:data-test" to run the test version of the update data script.

For the time being there are only placeholder files as these experiments are actually being done in an external 'experiments' folder in my local machine.

Watch this branch ;)